// src/schemas/expediente.schema.ts
export interface ExpedienteFormData {
  nombre: string;
  especie: string;
  raza: string;
  edad: string;
  sexo: string;
  peso: string;
  tamano: string;
  tipoMovimiento: string;
  fecha: string;
  motivo: string;
  lugar: string;
  descripcion: string;
  comportamientoAgresivo: boolean;
  enfermedadDegenerativa: boolean;
  discapacidades: boolean;
}